data-source/requirements.txt
kafka-python==1.3.5
six==1.10.0
